requirements.txt
certifi==2020.4.5.1
Pillow==7.1.2
